workload/all: relocate from pkg/ccl/workloadccl/allccl
The aggregator lived under pkg/ccl because it bundled the (then)
CCL-only roachmart workload; with roachmart relocated, the aggregator
follows. This deletes the last subdirectory of pkg/ccl/workloadccl;
the matching CODEOWNERS entry and a stale pkg/cli/cli.go comment are
repointed at the new path.

Epic: none
Release note: None
