+++ pkg/BUILD.bazel
@@ -68,7 +68,6 @@ ALL_TESTS = [
     "//pkg/ccl/testccl/sqlccl:sqlccl_test",
     "//pkg/ccl/testccl/sqlstatsccl:sqlstatsccl_test",
     "//pkg/ccl/testccl/workload/schemachange:schemachange_test",
-    "//pkg/ccl/workloadccl/allccl:allccl_test",
     "//pkg/cli/clierror:clierror_test",
     "//pkg/cli/clisqlclient:clisqlclient_test",
     "//pkg/cli/clisqlexec:clisqlexec_test",
@@ -900,6 +899,7 @@ ALL_TESTS = [
     "//pkg/util/vfsutil:vfsutil_test",
     "//pkg/util/yamlutil:yamlutil_test",
     "//pkg/util:util_test",
+    "//pkg/workload/all:all_test",
     "//pkg/workload/bank:bank_test",
     "//pkg/workload/cli:cli_test",
     "//pkg/workload/conflict:conflict_test",
@@ -1062,8 +1062,6 @@ GO_TARGETS = [
     "//pkg/ccl/testccl/sqlstatsccl:sqlstatsccl_test",
     "//pkg/ccl/testccl/workload/schemachange:schemachange_test",
     "//pkg/ccl/testutilsccl:testutilsccl",
-    "//pkg/ccl/workloadccl/allccl:allccl",
-    "//pkg/ccl/workloadccl/allccl:allccl_test",
     "//pkg/ccl:ccl",
     "//pkg/changefeed/changefeedpb:changefeedpb",
     "//pkg/changefeed/changefeedpb:checkpointpb",
@@ -3125,6 +3123,8 @@ GO_TARGETS = [
     "//pkg/util/yamlutil:yamlutil_test",
     "//pkg/util:util",
     "//pkg/util:util_test",
+    "//pkg/workload/all:all",
+    "//pkg/workload/all:all_test",
     "//pkg/workload/bank:bank",
     "//pkg/workload/bank:bank_test",
     "//pkg/workload/bulkingest:bulkingest",
