+++ pkg/cmd/workload/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/cockroachdb/cockroach/pkg/cmd/workload",
     visibility = ["//visibility:private"],
     deps = [
-        "//pkg/ccl/workloadccl/allccl",
+        "//pkg/workload/all",
         "//pkg/workload/cli",
         "//pkg/workload/cli/fixturescmd",
     ],
