+++ pkg/cmd/workload/main.go
@@ -8,7 +8,7 @@ package main
 import (
 	"os"
 
-	_ "github.com/cockroachdb/cockroach/pkg/ccl/workloadccl/allccl" // init hooks
+	_ "github.com/cockroachdb/cockroach/pkg/workload/all" // registers all workloads
 	workloadcli "github.com/cockroachdb/cockroach/pkg/workload/cli"
 	_ "github.com/cockroachdb/cockroach/pkg/workload/cli/fixturescmd" // registers fixtures command
 )
